# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Satara Development Internal Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /land-prospects:
    get:
      operationId: listLandProspects
      tags: [land-prospects]
      summary: List land prospects
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of land prospects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LandProspect"
    post:
      operationId: createLandProspect
      tags: [land-prospects]
      summary: Create land prospect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandProspectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandProspect"
  /land-prospects/{id}:
    get:
      operationId: getLandProspect
      tags: [land-prospects]
      summary: Get land prospect
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Land prospect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandProspect"
    patch:
      operationId: updateLandProspect
      tags: [land-prospects]
      summary: Update land prospect
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandProspectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandProspect"
...
  /legal:
    get:
      operationId: listLegalDocuments
      tags: [legal]
      summary: List legal documents
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalDocument"
    post:
      operationId: createLegalDocument
      tags: [legal]
      summary: Create legal document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalDocument"
  /legal/{id}:
    patch:
      operationId: updateLegalDocument
      tags: [legal]
      summary: Update legal document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalDocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalDocument"
  /legal/bankable-gate/{projectId}:
    get:
      operationId: getBankableGate
      tags: [legal]
      summary: Get bankable gate status for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bankable gate status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankableGate"

  # ─── LEADS ───────────────────────────────────────────────────
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/marketing-kpi:
    get:
      operationId: getMarketingKpi
      tags: [leads]
      summary: Marketing KPI metrics
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Marketing KPI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketingKpi"

  # ─── CUSTOMERS ───────────────────────────────────────────────
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: belowMinimum
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update material stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"

  # ─── HANDOVERS ───────────────────────────────────────────────
...
components:
  schemas:
...
    LandProspect:
      type: object
      required: [id, lokasi, luas, hargaM2, status, roi, margin, projectId, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: ["integer", "null"] }
        lokasi: { type: string }
        luas: { type: number }
        hargaM2: { type: number }
        status:
          type: string
          enum: [prospek_baru, survey, analisis_kompetitor, negosiasi, legal_checking, pks_mou, ditolak]
        roi: { type: number }
        margin: { type: number }
        aksesJalan: { type: ["number", "null"] }
        riskLevel:
          type: ["string", "null"]
          enum: [green, yellow, red, null]
        catatan: { type: ["string", "null"] }
        createdAt: { type: string }
    LandProspectInput:
      type: object
      required: [lokasi, luas, hargaM2, status]
      properties:
        projectId: { type: integer }
        lokasi: { type: string }
        luas: { type: number }
        hargaM2: { type: number }
        status:
          type: string
          enum: [prospek_baru, survey, analisis_kompetitor, negosiasi, legal_checking, pks_mou, ditolak]
        roi: { type: number }
        margin: { type: number }
        aksesJalan: { type: number }
        catatan: { type: string }
    LandProspectUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [prospek_baru, survey, analisis_kompetitor, negosiasi, legal_checking, pks_mou, ditolak]
        roi: { type: number }
        margin: { type: number }
        aksesJalan: { type: number }
        catatan: { type: string }
...
    LegalDocument:
      type: object
      required: [id, projectId, tipeDokumen, status, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        tipeDokumen:
          type: string
          enum: [SHM, AJB, balik_nama, split, PBB, PKKPR, SPPL, PBG, SLF, bank_ready, sikumbang, lainnya]
        status:
          type: string
          enum: [pending, in_progress, approved, rejected]
        pic: { type: ["string", "null"] }
        expiry: { type: ["string", "null"] }
        catatan: { type: ["string", "null"] }
        createdAt: { type: string }
    LegalDocumentInput:
      type: object
      required: [projectId, tipeDokumen, status]
      properties:
        projectId: { type: integer }
        tipeDokumen:
          type: string
          enum: [SHM, AJB, balik_nama, split, PBB, PKKPR, SPPL, PBG, SLF, bank_ready, sikumbang, lainnya]
        status:
          type: string
          enum: [pending, in_progress, approved, rejected]
        pic: { type: string }
        expiry: { type: string }
        catatan: { type: string }
    LegalDocumentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, approved, rejected]
        pic: { type: string }
        expiry: { type: string }
        catatan: { type: string }
    BankableGate:
      type: object
      required: [projectId, isBankable, checks]
      properties:
        projectId: { type: integer }
        isBankable: { type: boolean }
        checks:
          type: array
          items:
            type: object
            required: [name, status]
            properties:
              name: { type: string }
              status:
                type: string
                enum: [green, yellow, red]
              notes: { type: ["string", "null"] }

    # ─── LEADS ───────────────────────────────────────────────────
    Lead:
      type: object
      required: [id, nama, kontak, status, source, projectId, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        nama: { type: string }
        kontak: { type: string }
        source:
          type: string
          enum: [meta_ads, instagram, tiktok, facebook, broker, afiliator, event, referral, lainnya]
        status:
          type: string
          enum: [lead_masuk, qualified, survey, booking, berkas_kpr, akad, batal]
        assignedTo: { type: ["string", "null"] }
        campaign: { type: ["string", "null"] }
        followUpAt: { type: ["string", "null"] }
        alasanBatal: { type: ["string", "null"] }
        createdAt: { type: string }
    LeadInput:
      type: object
      required: [projectId, nama, kontak, source, status]
      properties:
        projectId: { type: integer }
        nama: { type: string }
        kontak: { type: string }
        source:
          type: string
          enum: [meta_ads, instagram, tiktok, facebook, broker, afiliator, event, referral, lainnya]
        status:
          type: string
          enum: [lead_masuk, qualified, survey, booking, berkas_kpr, akad, batal]
        assignedTo: { type: string }
        campaign: { type: string }
    LeadUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [lead_masuk, qualified, survey, booking, berkas_kpr, akad, batal]
        assignedTo: { type: string }
        followUpAt: { type: string }
        alasanBatal: { type: string }
    MarketingKpi:
      type: object
      required: [totalLeads, cpl, conversionRate, bookingToAkadRate, surveyRate, responseTimeAvg]
      properties:
        totalLeads: { type: integer }
        cpl: { type: number }
        conversionRate: { type: number }
        bookingToAkadRate: { type: number }
        surveyRate: { type: number }
        responseTimeAvg: { type: number }
        bySource:
          type: array
          items:
            type: object
            required: [source, count]
            properties:
              source: { type: string }
              count: { type: integer }

    # ─── CUSTOMERS ───────────────────────────────────────────────
...
    Material:
      type: object
      required: [id, projectId, item, stok, satuan, minimumStock, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        item: { type: string }
        stok: { type: number }
        satuan: { type: string }
        vendor: { type: ["string", "null"] }
        harga: { type: ["number", "null"] }
        minimumStock: { type: number }
        isBelowMinimum: { type: boolean }
        createdAt: { type: string }
    MaterialInput:
      type: object
      required: [projectId, item, stok, satuan, minimumStock]
      properties:
        projectId: { type: integer }
        item: { type: string }
        stok: { type: number }
        satuan: { type: string }
        vendor: { type: string }
        harga: { type: number }
        minimumStock: { type: number }
    MaterialUpdate:
      type: object
      properties:
        stok: { type: number }
        vendor: { type: string }
        harga: { type: number }
        minimumStock: { type: number }

    # ─── HANDOVER ─────────────────────────────────────────────────
...